refactor(#6): Hook for getting the current user position
